Add test Polaris *.xvg output file.
Generated a test Polaris output file that contains the header information of
the calculation, and the result of two snapshot calculations. Note that the
first two snapshots are still within the burn-in period of the simulation, and
are therefore physically unrealistic. This test file is simply used to test
file parsing/conversion programs to ensure that they conform to official
Polaris output.