Add *.sle extraction of MD box size into *.h5 file.
Add a table in the MD simulation parameters table to hold the (x,y,z) box size
for the simulation. This information is read off of the *.sle file. Currently
the parsing is hardcoded, and it is expected that a *.sle file with the same
filename is in the directory with the *.xvg file.